+++ public/app/features/search/page/components/FolderSection.tsx
@@ -4,7 +4,7 @@ import { useAsync, useLocalStorage } from 'react-use';
 
 import { GrafanaTheme } from '@grafana/data';
 import { getBackendSrv } from '@grafana/runtime';
-import { Checkbox, CollapsableSection, Icon, stylesFactory, useTheme } from '@grafana/ui';
+import { Card, Checkbox, CollapsableSection, Icon, Spinner, stylesFactory, useTheme } from '@grafana/ui';
 import impressionSrv from 'app/core/services/impression_srv';
 import { getSectionStorageKey } from 'app/features/search/utils';
 import { useUniqueId } from 'app/plugins/datasource/influxdb/components/useUniqueId';
@@ -128,7 +128,15 @@ export const FolderSection: FC<SectionHeaderProps> = ({
 
   const renderResults = () => {
     if (!results.value?.length) {
-      return <div>No items found</div>;
+      if (results.loading) {
+        return <Spinner />;
+      }
+
+      return (
+        <Card>
+          <Card.Heading>No results found</Card.Heading>
+        </Card>
+      );
     }
 
     return results.value.map((v) => {
